// typings.d.ts
export interface Post{
    _id:string;
    title:string;
    description:string;
    _createdAt:string;
    author:{
        name:string;
        image:string;
    };
    mainImage:{
        asset:{
            url:string;
        };
    };
    slug:{
        current:string;
    };
    body: [object];

};
export interface PostList{
    _id:string;
    title:string;
    description:string;
    comments:Comment[];
    _createdAt:string;
    author:{
        name:string;
        image:string;
    };
    mainImage:{
        asset:{
            url:string;
        };
    };
    slug:{
        current:string;
    };
    body: [object];

};

export interface Comment{
    approved:boolean;
    comment:string;
    email:string;
    name:string;
    post:{
        _ref:string;
        _type:string;
    };
    _createdAt:string;
    _id:string;
    _rev:string;
    _type:string;
    _updatedAt:string;
};